Add FlatGroupByHash cached mode
Adds support for caching the computed hash values in FlatGroupByHash
hash tables when the GroupByHash may be spilled or when the hash
calculation is expensive to compute.

In the case of spillable hash tables, spilling will involve sorting the
table contents based on the raw hash value- which can be extremely
expensive if the value is being recomputed on each comparison compared
to simply accessed from the table.

Similarly, re-hashing large FlatGroupByHash instances can be
significantly more expensive without the hash value cached when there
are a large number of columns or when the hash calculation itself is
relatively expensive.