Add reduced size VariableWidthData
Introduces AppendOnlyVariableWidthData, which is related to the
VariableWidthData implementation except that the fixed size pointer
saves 4 bytes per entry as a result of not needing to encode the
variable width data length for each entry. This comes at the cost of not
supporting the ability to relocate or free individual entries, but
structures not needing that functionality can see an appreciable memory
usage improvement by using the new implementation instead.

This commit also updates FlatHash to use the new append only variation
to reduce the fixed record size per entry.